requirements.txt
python-dotenv==1.0.0
google-generativeai==0.3.2
web3==6.15.1
py-solc-x==1.1.1
google-cloud-aiplatform==1.42.1 